Make WINDOW_SIZE an argument in keywords.
Update keywords.py to make WINDOW_SIZE an argument in keywords, so the ngram length for keywords can be specified in calls to keywords.py. Previously, WINDOW_SIZE was hard-coded with no option to replace it, so the only way to change it was to change the hard-coded value.

Change implemented by Nancy McCracken.